third_party/WebKit/LayoutTests/fast/dom/shadow/style-with-shadow-pseudo-element-expected.txt

CONSOLE WARNING: ::shadow pseudo-element is deprecated. See https://www.chromestatus.com/features/6750456638341120 for more details.
CONSOLE WARNING: Calling Element.createShadowRoot() for an element which already hosts a shadow root is deprecated. See https://www.chromestatus.com/features/4668884095336448 for more details.
Test for ::shadow, http://crbug.com/309504.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS borderColorOf(getNodeInComposedTree("host/top-div")) is "rgb(0, 128, 0)"
PASS borderColorOf(getNodeInComposedTree("host/top-span")) is not "rgb(0, 128, 0)"
PASS borderColorOf(getNodeInComposedTree("host/not-top-div")) is not "rgb(0, 128, 0)"
PASS borderColorOf(getNodeInComposedTree("host/not-top-span")) is not "rgb(0, 128, 0)"
PASS borderColorOf(getNodeInComposedTree("host/span1")) is not "rgb(0, 128, 0)"
PASS borderColorOf(getNodeInComposedTree("host/host1/span2")) is not "rgb(0, 128, 0)"
PASS borderColorOf(getNodeInComposedTree("host/host1/host2/span3")) is "rgb(0, 128, 0)"
PASS borderColorOf(getNodeInComposedTree("host/host1/host2/host3/span4")) is not "rgb(0, 128, 0)"
PASS borderColorOf(getNodeInComposedTree("host/target")) is "rgb(0, 128, 0)"
PASS borderColorOf(getNodeInComposedTree("host/target")) is "rgb(0, 128, 0)"
PASS borderColorOf(getNodeInComposedTree("host/target")) is "rgb(255, 0, 0)"
PASS borderColorOf(getNodeInComposedTree("host/target")) is "rgb(0, 128, 0)"
PASS borderColorOf(getNodeInComposedTree("host/target")) is "rgb(0, 128, 0)"
PASS borderColorOf(getNodeInComposedTree("host/target")) is "rgb(0, 128, 0)"
PASS borderColorOf(getNodeInComposedTree("host//target")) is "rgb(0, 128, 0)"
PASS borderColorOf(getNodeInComposedTree("host//target")) is "rgb(0, 0, 0)"
PASS borderColorOf(getNodeInComposedTree("host/host2/target")) is "rgb(0, 128, 0)"
PASS borderColorOf(getNodeInComposedTree("host/target")) is "rgb(0, 0, 0)"
PASS borderColorOf(getNodeInComposedTree("host/host2/target")) is "rgb(0, 0, 0)"
PASS borderColorOf(getNodeInComposedTree("host/host2/target")) is "rgb(0, 128, 0)"
PASS borderColorOf(getNodeInComposedTree("host/host2/target")) is "rgb(0, 0, 0)"
PASS borderColorOf(getNodeInComposedTree("host/x-bar-host/x-zot-host/x-zot-target")) is not "rgb(255, 0, 0)"
PASS successfullyParsed is true

TEST COMPLETE
